+++ tests/testthat/test-rows.R
@@ -154,6 +154,34 @@ test_that("`rows_insert()` with `in_place = TRUE` and `returning`", {
   expect_equal(df_inserted %>% collect(), tibble(x = 1:4, y = c(11:13, 24L)))
 })
 
+test_that("rows_get_or_execute() gives error context", {
+  con <- DBI::dbConnect(RSQLite::SQLite(), ":memory:")
+  on.exit(DBI::dbDisconnect(con))
+  DBI::dbWriteTable(con, "mtcars", tibble(x = 1, y = 1), overwrite = TRUE, temporary = TRUE)
+  DBI::dbExecute(con, "CREATE UNIQUE INDEX `mtcars_x` ON `mtcars` (`x`)")
+
+  expect_snapshot({
+    (expect_error(
+      rows_append(
+        tbl(con, "mtcars"),
+        tibble(x = 1),
+        copy = TRUE,
+        in_place = TRUE
+      )
+    ))
+
+    (expect_error(
+      rows_append(
+        tbl(con, "mtcars"),
+        tibble(x = 1),
+        copy = TRUE,
+        in_place = TRUE,
+        returning = x
+      )
+    ))
+  }, transform = snap_transform_dbi)
+})
+
 test_that("`sql_query_insert()` works", {
   df_y <- lazy_frame(
     a = 2:3, b = c(12L, 13L), c = -(2:3), d = c("y", "z"),